requirements.txt
pandas
streamlit
numpy
matplotlib
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
scikit-learn==1.0.2
scipy==1.7.3
sklearn==0.0
SQLAlchemy==1.4.29
threadpoolctl==3.0.0
Werkzeug==2.0.2
seaborn
prophet